+++ apps/ai-dial-admin/src/components/Publications/Assets/Prompt/PromptsProperties.tsx
