+++ apps/ai-dial-admin/src/components/Publications/Assets/Prompt/tests/PromptsProperties.spec.tsx
