+++ apps/ai-dial-admin/src/components/Publications/View/Properties.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { ApplicationRoute } from '@/src/types/routes';
 import { Publication } from '@/src/models/dial/publications';
 import BasePublicationProperties from '@/src/components/Publications/Properties/Properties';
-import PromptsPropertiesList from '@/src/components/Publications/Assets/Files/Prompt/PromptsPropertiesList';
+import PromptsPropertiesList from '@/src/components/Publications/Assets/Prompt/PromptsPropertiesList';
 import FilesProperties from '@/src/components/Publications/Assets/Files/FilesProperties';
 import ApplicationProperties from '@/src/components/Publications/Assets/Application/ApplicationProperties';
 import { DialApplicationScheme } from '@/src/models/dial/application';
